+++ src/lib/primitives/form-group-node.ts
@@ -15,6 +15,7 @@ import { createNodeMetadata } from '../metadata/create-node-metadata';
 import { REQUIRED_METADATA } from '../validation/validators/required';
 import { createCallableNodeApi } from './utils/create-callable-node-api';
 import { createNodeValueSignal } from './utils/create-node-value-signal';
+import { createNodeErrorsSignal } from './utils/create-node-errors-signal';
 import { registerNodeInputConfig } from '../configuration/node-input-config';
 import { markAsFieldContext } from '../validation/utils/field-context-marker';
 import { createAsyncValidation } from '../validation/create-async-validation';
@@ -578,7 +579,7 @@ export class FormGroupNode<TNodes extends Nodes> {
       resetToInitial: () => this.resetToInitial(),
       validators: createValidatorQuery(this.validators.asReadonly(), () => this.validatorResolution().resolvedValidators),
       setValidators: (next: ValidatorSource<FormValue<TNodes>>) => this.setValidators(next),
-      errors: this.errors,
+      errors: createNodeErrorsSignal(this.errors, this.allErrors),
       allErrors: this.allErrors,
       valid: this.valid,
       invalid: this.invalid,